[FEAT] Enable Seen on X
